Rename files to markdown_editor
