+++ index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <title>Markdown Editor</title>
-    <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="markdown_editor.css" />
     <link rel="stylesheet" href="syntax_highlight.css" />
   </head>
   <body>
@@ -19,7 +19,7 @@
         <div id="preview"></div>
       </div>
     </div>
-    <script src="app.js"></script>
+    <script src="markdown_editor.js"></script>
     <script src="codeBlockSyntax_java.js"></script>
   </body>
 </html>
